refactor(buzz): extract shouldAbortPressOnLeave to pure util file
Round-4 code-review audit follow-up for PR #2307.

`shouldAbortPressOnLeave` (and the `PRESS_MOVE_TOLERANCE` constant it
defaults from) were exported from InteractiveTipBuzzButton.tsx, so the
unit test for this pure helper transitively loaded the whole Mantine
component graph. Moved both into a sibling pure util file
(InteractiveTipBuzzButton.utils.ts — no React/Mantine imports).

- component imports `shouldAbortPressOnLeave` from the util file
- test imports `shouldAbortPressOnLeave` from the util file
- pure move/refactor only — behavior identical, same exported symbol
